+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import type { AppProps } from "next/app";
+import { AppProps } from "next/app";
 import { RecoilRoot } from "recoil";
 import "../styles/globals.css";
